Refs #33355 -- Moved SQLite functions to separate module.
